requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coverage==5.4
dj-database-url==0.5.0
Django==3.1.6
django-rest-framework==0.1.0
djangorestframework==3.12.2
flake8==3.8.4
flake8-django==1.1.1
gunicorn==20.0.4
httpretty==1.0.5
idna==2.10
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2021.1
requests==2.25.1
responses==0.12.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
